Move run_hooks helper to a mixin and use it for EditView hooks
